combined interface and implementation class in ingest_data.py and renamed "pipeline_classes" to "classes_for_steps" and deleted unnecessary "return None"
